feat: refactor styles and config
