+++ components/Layout/index.tsx
@@ -31,41 +31,51 @@ const spin = keyframes`0% {
 }`
 
 const RadialBackground = styled.div<RadialBackgroundProps>`
-&:hover {
-  color: rgb(88 199 250 / 100%);
-  transition: color 1s;
-}
-&:hover:before, &:hover:after {
-  animation: none;
-  opacity: 0;
-}
+  @property --rotate {
+    syntax: "<angle>";
+    initial-value: 132deg;
+    inherits: false;
+  }
+    
+  --radiusGradient-height: 50vh;
+  --radiusGradient-width: calc(var(--radiusGradient-height) / 1.5);
 
-&::before {
-  content: "";
-  border-radius: 8px;
-  background-image: linear-gradient(var(--rotate), ${props => props.primary}, ${props => props.secondary} 43%, ${props => props.tertiary});
+  &:hover {
+    color: rgb(88 199 250 / 100%);
+    transition: color 1s;
+  }
+  &:hover:before, &:hover:after {
+    animation: none;
+    opacity: 0;
+  }
+
+  &::before {
+    content: "";
+    border-radius: 8px;
+    background-image: linear-gradient(var(--rotate), ${props => props.primary}, ${props => props.secondary} 43%, ${props => props.tertiary});
     position: absolute;
     z-index: -1;
     animation: spin 2.5s linear infinite;
-}
+  }
 
-&::after {
-  position: absolute;
-  content: "";
-  top: calc(var(--radiusGradient-height) / 6);
-  left: 0;
-  right: 0;
-  z-index: -1;
-  height: 100%;
-  width: 100%;
-  margin: 0 auto;
-  transform: scale(0.8);
-  filter: blur(calc(var(--radiusGradient-height) / 6));
-  background-image: linear-gradient(var(--rotate), ${props => props.primary}, ${props => props.secondary} 43%, ${props => props.tertiary});
-  opacity: 0.3;
-  transition: opacity .5s;
-  animation: ${spin} 8.5s linear infinite;
-`
+  &::after {
+    position: absolute;
+    content: "";
+    top: calc(var(--radiusGradient-height) / 6);
+    left: 0;
+    right: 0;
+    z-index: -1;
+    height: 100%;
+    width: 100%;
+    margin: 0 auto;
+    transform: scale(0.8);
+    filter: blur(calc(var(--radiusGradient-height) / 6));
+    background-image: linear-gradient(var(--rotate), ${props => props.primary}, ${props => props.secondary} 43%, ${props => props.tertiary});
+    opacity: 0.3;
+    transition: opacity .5s;
+    animation: ${spin} 8.5s linear infinite;
+  }
+`;
 
 const Layout = ({ children, title, description, colors, radialBackground: isRadialBackground }: Props) => {
   return (
