+++ pages/_document.tsx
@@ -1,14 +1,14 @@
 import { Html, Head, Main, NextScript } from 'next/document'
-import { Config } from '../components/Config'
- 
+import { Config } from '../Config'
+
 export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <body style={{ background:Config.default.background, color: Config.default.foreground }}>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
+    return (
+        <Html lang="en">
+            <Head />
+            <body style={{ background: Config.default.background, color: Config.default.foreground }}>
+                <Main />
+                <NextScript />
+            </body>
+        </Html>
+    )
 }
